QA-Automation-Process/requirements_win.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cycler==0.12.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
imageio==2.35.1
imgaug==0.4.0
keras==2.13.1
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.4.1
mss==9.0.2
namex==0.0.8
networkx==3.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
PyGetWindow==0.0.9
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
python-dateutil==2.9.0.post0
pywin32==306
pywinauto==0.6.8
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.1
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.5.0
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.0.4
wrapt==1.16.0
